Add tests for the new code quality test
